#747 Creation of diffCommon.py
